requirements/base.txt
pytz==2019.3  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==7.0.0  # https://github.com/python-pillow/Pillow
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.0.1  # https://github.com/evansd/whitenoise
redis==3.3.11  # https://github.com/antirez/redis
celery==4.4.0  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==1.5.0  # https://github.com/celery/django-celery-beat

# Django
# ------------------------------------------------------------------------------
django==2.2.9  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.8.1  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.11.0  # https://github.com/niwinz/django-redis
django-des # https://github.com/jamiecounsell/django-des
django-cors-headers # https://github.com/adamchainz/django-cors-headers
django-rest-swagger # https://github.com/marcgibbons/django-rest-swagger

# Django REST Framework
djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
djangorestframework-jwt # https://jpadilla.github.io/django-rest-framework-jwt/
#django-use-email-as-username # https://github.com/jmfederico/django-use-email-as-username

# Para validação e formatação de CNPJ e CPF
# https://github.com/poliquin/brazilnum
brazilnum==0.8.8

# Para login com e-mail
# https://github.com/jmfederico/django-use-email-as-username
django-use-email-as-username==1.1.0

# Para filtros no DRF
# https://django-filter.readthedocs.io/en/master/#
django-filter==2.2.0
